Move InputGroup and SpacedGroup to separate components
