--- packages/flame/src/Groups/index.tsx
@@ -1,3 +1,0 @@
-import { InputGroup, SpacedGroup, InputGroupAddon, BeThereOrBeSquare } from './Groups';
-
-export { InputGroup, SpacedGroup, InputGroupAddon, BeThereOrBeSquare };
