+++ packages/flame/src/InputGroup/InputGroup.test.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import { createComponent } from 'test-utils';
-import { InputGroup } from './Groups';
+import { InputGroup } from './InputGroup';
 import { Input } from '../Input';
 import { Button } from '../Button';
 
-describe('Group', () => {
+describe('InputGroup', () => {
   describe('Snaphots', () => {
     it('should adjust border radii', () => {
       const component = createComponent(
