+++ packages/flame/src/InputGroup/index.tsx
@@ -1,0 +1,3 @@
+import { InputGroup, InputGroupAddon, BeThereOrBeSquare } from './InputGroup';
+
+export { InputGroup, InputGroupAddon, BeThereOrBeSquare };
