+++ packages/flame/src/SpacedGroup/index.tsx
@@ -1,0 +1,3 @@
+import { SpacedGroup } from './SpacedGroup';
+
+export { SpacedGroup };
